@W-18597952 - Scroll bar implementation for bonus products in the modal (#2689)
* Scrolling and Refactored code to extract bonus product item sub-component
